fix(ci): restore GitManager test line-limit compliance — split helper tests out of the oversized module
