+++ tests/unit/node/test_git_manager_environment.py
@@ -1,0 +1,121 @@
+"""GitManager environment and writable-worktree helper tests."""
+
+from __future__ import annotations
+
+import os
+from pathlib import Path
+
+import pytest
+
+from awf.node import git_manager as git_module
+from awf.node.git_manager import GitManager
+
+
+class TestGitEnvironment:
+    @pytest.mark.unit
+    async def test_run_uses_configured_environment(self, tmp_path: Path) -> None:
+        home = tmp_path / "home"
+        home.mkdir()
+        manager = GitManager(tmp_path / "work", env={"HOME": str(home), "AWF_TEST_ENV": "ok"})
+
+        result = await manager._run(  # noqa: SLF001 - narrow regression for subprocess env.
+            ["sh", "-c", 'printf \'%s:%s\' "$HOME" "$AWF_TEST_ENV"'],
+            operation="env",
+        )
+
+        assert result.stdout == f"{home}:ok"
+
+
+class TestAgentWritableWorktreeHelpers:
+    @pytest.mark.unit
+    async def test_prepare_agent_writable_worktree_skips_chown_when_not_root(
+        self,
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        monkeypatch.setattr(os, "geteuid", lambda: 1000)
+
+        async def _unexpected_to_thread(*_args: object, **_kwargs: object) -> None:
+            raise AssertionError("non-root process must not chown worktrees")
+
+        monkeypatch.setattr(git_module.asyncio, "to_thread", _unexpected_to_thread)
+        manager = GitManager(
+            tmp_path / "awf-work",
+            worktree_owner_uid=1000,
+            worktree_owner_gid=1000,
+        )
+
+        await manager._prepare_agent_writable_worktree(  # noqa: SLF001
+            layout_mirror=tmp_path / "mirror.git",
+            worktree_path=tmp_path / "worktree",
+        )
+
+    @pytest.mark.unit
+    def test_linked_worktree_git_dir_handles_absent_unreadable_and_relative_gitfiles(
+        self,
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        missing_gitfile = tmp_path / "missing"
+        missing_gitfile.mkdir()
+        assert git_module.linked_worktree_git_dir(missing_gitfile) is None
+
+        unreadable = tmp_path / "unreadable"
+        unreadable.mkdir()
+        unreadable_gitfile = unreadable / ".git"
+        unreadable_gitfile.write_text("gitdir: ../real.git\n", encoding="utf-8")
+        original_read_text = Path.read_text
+
+        def _read_text(path: Path, *args: object, **kwargs: object) -> str:
+            if path == unreadable_gitfile:
+                raise OSError("cannot read gitfile")
+            return original_read_text(path, *args, **kwargs)
+
+        monkeypatch.setattr(Path, "read_text", _read_text)
+        assert git_module.linked_worktree_git_dir(unreadable) is None
+
+        malformed = tmp_path / "malformed"
+        malformed.mkdir()
+        (malformed / ".git").write_text("not a gitdir pointer\n", encoding="utf-8")
+        assert git_module.linked_worktree_git_dir(malformed) is None
+
+        relative = tmp_path / "relative"
+        relative.mkdir()
+        (relative / ".git").write_text(
+            "gitdir: ../mirror.git/worktrees/ws_relative\n",
+            encoding="utf-8",
+        )
+
+        assert (
+            git_module.linked_worktree_git_dir(relative)
+            == (relative / "../mirror.git/worktrees/ws_relative").resolve()
+        )
+
+    @pytest.mark.unit
+    def test_chown_targets_skip_missing_and_duplicate_paths(
+        self,
+        tmp_path: Path,
+        monkeypatch: pytest.MonkeyPatch,
+    ) -> None:
+        chowned: list[tuple[Path, int, int]] = []
+        file_path = tmp_path / "owned-file"
+        file_path.write_text("content\n", encoding="utf-8")
+        missing_path = tmp_path / "missing"
+
+        monkeypatch.setattr(
+            os,
+            "chown",
+            lambda path, uid, gid: chowned.append((Path(path), uid, gid)),
+        )
+
+        git_module._chown_targets(  # noqa: SLF001
+            (
+                git_module._ChownTarget(file_path, recursive=True),  # noqa: SLF001
+                git_module._ChownTarget(file_path, recursive=True),  # noqa: SLF001
+                git_module._ChownTarget(missing_path, recursive=False),  # noqa: SLF001
+            ),
+            1000,
+            1001,
+        )
+
+        assert chowned == [(file_path, 1000, 1001)]
